change python wrapper building process
Generate wrapper with cmake custom command, and build wrapper as
python extension with distutils.

If we build wrapper as static library and link it to esn shared
library, we need to add target_link_libraries() to the main cmake
script, which is out of scope of the python wrapper source directory.

If we build wrapper as shared library with cmake, setup.py doesn't
provide a convinient way to install this shared library.

If we generate wrapper with python distutils, python may use source
directory for keeping built files.